+++ scripts/merge_results.py
@@ -43,8 +43,21 @@ cols = set(result.columns.tolist())
 cols_include = ['sequences', 'qlen', 'length', 'pident', 'mismatch', 'qstart', 'qend', 'sstart', 'send', 'gaps', 'evalue', 'taxonomy']
 cols_all = cols_include + list(cols - set(cols_include))
 
+result2 = result.copy(deep=True)
+result3 = result.copy(deep=True)
 result = result[cols_all]
 result.to_csv(snakemake.output["complete"], index_label="seqid")
+
 # filter out OTUs without taxonomic annotation
 result = result.loc[~nas, :]
 result.to_csv(snakemake.output["filtered"], index_label="seqid")
+
+# save otu table
+samples = list(cols - set(cols_include))
+
+result2 = result2[samples]
+result2.to_csv(snakemake.output["otus"], index_label="seqid")
+
+# save metadata table
+result3 = result3[cols_include]
+result3.to_csv(snakemake.output["metadata"], index_label="seqid")